Include few cutsom fields in user entity, use long instead of integer for user id

+++ user-service/src/main/java/com/project/userservice/services/UserService.java
@@ -52,7 +52,7 @@ public interface UserService {
      * @param userId The ID of the user to retrieve.
      * @return An optional containing the user if found.
      */
-    Optional<User> getUserById(Integer userId);
+    Optional<User> getUserById(Long userId);
 
     /**
      * Creates a new user.
